ULEB128 refactoring
