+++ src/Parquet/Extensions/EncodingExtensions.cs
@@ -33,61 +33,75 @@ namespace Parquet.Extensions {
 
         #region [ ZigZag ]
 
-        public static uint ZigZagEncode(this int num) {
-            return (uint)((num >> 31) ^ (num << 1));
+        public static long ZigZagEncode(this long num) {
+            return (num >> 63) ^ (num << 1);
         }
 
-
-        public static ulong ZigZagEncode(this long num) {
-            return (ulong)((num >> 63) ^ (num << 1));
+        public static long ZigZagDecode(this long raw) {
+            long temp = (((raw << 63) >> 63) ^ raw) >> 1;
+            return temp ^ (raw & (1L << 63));
         }
 
+        public static long ReadZigZagVarLong(this Span<byte> data, ref int offset) {
+            long raw = data.ReadUnsignedVarLong(ref offset);
+            return ZigZagDecode(raw);
+        }
 
         #endregion
 
         #region [ Variable Length Encoding ]
 
+        /// <summary>
+        /// Encodes in ULEB128 and returns number of emitted bytes.
+        /// https://en.wikipedia.org/wiki/LEB128#Unsigned_LEB128
+        /// </summary>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static void WriteUnsignedVarInt(this int value, Span<byte> dest, ref int consumed) {
+        public static int ULEB128Encode(this ulong value, Span<byte> dest) {
+            int i = 0;
             while(value > 127) {
                 byte b = (byte)((value & 0x7F) | 0x80);
-
-                dest[consumed++] = b;
-
+                dest[i++] = b;
                 value >>= 7;
             }
-
-            dest[consumed++] = (byte)value;
+            dest[i++] = (byte)value;
+            return i;
         }
 
-        public static int WriteUnsignedVarInt(this Stream destination, int value) {
-            int consumed = 0;
-            Span<byte> buf = stackalloc byte[4];
-            value.WriteUnsignedVarInt(buf, ref consumed);
-            destination.WriteSpan(buf.Slice(0, consumed));
-            return consumed;
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static void WriteULEB128(this Stream destination, ulong value) {
+            Span<byte> buf = stackalloc byte[10];   // max 10 bytes
+            int size = value.ULEB128Encode(buf);
+            destination.WriteSpan(buf.Slice(0, size));
         }
 
-        public static int WriteUnsignedVarLong(this Stream destination, ulong value) {
-            int consumed = 0;
-            Span<byte> buf = stackalloc byte[8];
-            while(value > 127) {
-                byte b = (byte)((value & 0x7F) | 0x80);
+        /// <summary>
+        /// Read a value using the unsigned, variable int encoding.
+        /// </summary>
+        public static int ReadUnsignedVarInt(this Span<byte> data, ref int offset) {
+            int result = 0;
+            int shift = 0;
 
-                buf[consumed++] = b;
-
-                value >>= 7;
+            while(true) {
+                byte b = data[offset++];
+                result |= ((b & 0x7F) << shift);
+                if((b & 0x80) == 0)
+                    break;
+                shift += 7;
             }
 
-            buf[consumed++] = (byte)value;
-#if NETSTANDARD2_0
-            destination.Write(buf.Slice(0, consumed).ToArray(), 0, consumed);
-#else
-            destination.Write(buf.Slice(0, consumed));
-#endif
-            return consumed;
+            return result;
         }
 
+        public static long ReadUnsignedVarLong(this Span<byte> data, ref int offset) {
+            long value = 0;
+            int i = 0;
+            long b;
+            while(((b = data[offset++]) & 0x80) != 0) {
+                value |= (b & 0x7F) << i;
+                i += 7;
+            }
+            return value | (b << i);
+        }
 
         #endregion
